pyproject.toml
[project]
name = "pex"
version = "0.1.0"
description = "Expenses tracker application."
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["typer", "python-dateutil" ]

[project.scripts]
pex = "main:app"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]


